+++ atl-checker/src/edg/atledg/mod.rs
@@ -359,213 +359,3 @@ impl<G: GameStructure> ExtendedDependencyGraph<ATLVertex> for ATLDependencyGraph
         }
     }
 }
-
-#[cfg(test)]
-mod test {
-    use std::collections::HashSet;
-    use std::sync::Arc;
-
-    use crate::edg::atledg::pmoves::{
-        DeltaIterator, PartialMoveChoice, PartialMoveIterator, PmovesIterator,
-    };
-    use crate::game_structure::{DynVec, EagerGameStructure};
-
-    #[test]
-    fn partial_move_iterator_01() {
-        let partial_move = vec![
-            PartialMoveChoice::RANGE(2),
-            PartialMoveChoice::SPECIFIC(1),
-            PartialMoveChoice::RANGE(2),
-        ];
-
-        let mut iter = PartialMoveIterator::new(&partial_move);
-        assert_eq!(iter.next(), Some(vec![0, 1, 0]));
-        assert_eq!(iter.next(), Some(vec![0, 1, 1]));
-        assert_eq!(iter.next(), Some(vec![1, 1, 0]));
-        assert_eq!(iter.next(), Some(vec![1, 1, 1]));
-        assert_eq!(iter.next(), None);
-    }
-
-    #[test]
-    fn vars_iterator_01() {
-        let moves = vec![2, 3, 2];
-        let mut players = HashSet::new();
-        players.insert(0);
-        players.insert(2);
-        let mut iter = PmovesIterator::new(moves, players);
-
-        assert_eq!(
-            &iter.next(),
-            &Some(vec![
-                PartialMoveChoice::SPECIFIC(0),
-                PartialMoveChoice::RANGE(3),
-                PartialMoveChoice::SPECIFIC(0)
-            ])
-        );
-        assert_eq!(
-            &iter.next(),
-            &Some(vec![
-                PartialMoveChoice::SPECIFIC(1),
-                PartialMoveChoice::RANGE(3),
-                PartialMoveChoice::SPECIFIC(0)
-            ])
-        );
-        assert_eq!(
-            &iter.next(),
-            &Some(vec![
-                PartialMoveChoice::SPECIFIC(0),
-                PartialMoveChoice::RANGE(3),
-                PartialMoveChoice::SPECIFIC(1)
-            ])
-        );
-        assert_eq!(
-            &iter.next(),
-            &Some(vec![
-                PartialMoveChoice::SPECIFIC(1),
-                PartialMoveChoice::RANGE(3),
-                PartialMoveChoice::SPECIFIC(1)
-            ])
-        );
-    }
-
-    #[test]
-    fn vars_iterator_02() {
-        let mut players = HashSet::new();
-        players.insert(2);
-        let mut iter = PmovesIterator::new(vec![2, 3, 3], players);
-
-        let value = iter.next().unwrap();
-        assert_eq!(value[0], PartialMoveChoice::RANGE(2));
-        assert_eq!(value[1], PartialMoveChoice::RANGE(3));
-        assert_eq!(value[2], PartialMoveChoice::SPECIFIC(0));
-
-        let value = iter.next().unwrap();
-        assert_eq!(value[0], PartialMoveChoice::RANGE(2));
-        assert_eq!(value[1], PartialMoveChoice::RANGE(3));
-        assert_eq!(value[2], PartialMoveChoice::SPECIFIC(1));
-
-        let value = iter.next().unwrap();
-        assert_eq!(value[0], PartialMoveChoice::RANGE(2));
-        assert_eq!(value[1], PartialMoveChoice::RANGE(3));
-        assert_eq!(value[2], PartialMoveChoice::SPECIFIC(2));
-
-        let value = iter.next();
-        assert_eq!(value, None);
-    }
-
-    #[test]
-    fn vars_iterator_03() {
-        // Both players choose. So we should end up with every move vector
-        let mut players = HashSet::new();
-        players.insert(0);
-        players.insert(1);
-        let mut iter = PmovesIterator::new(vec![3, 3], players);
-
-        let value = iter.next().unwrap();
-        assert_eq!(value[0], PartialMoveChoice::SPECIFIC(0));
-        assert_eq!(value[1], PartialMoveChoice::SPECIFIC(0));
-
-        let value = iter.next().unwrap();
-        assert_eq!(value[0], PartialMoveChoice::SPECIFIC(1));
-        assert_eq!(value[1], PartialMoveChoice::SPECIFIC(0));
-
-        let value = iter.next().unwrap();
-        assert_eq!(value[0], PartialMoveChoice::SPECIFIC(2));
-        assert_eq!(value[1], PartialMoveChoice::SPECIFIC(0));
-
-        let value = iter.next().unwrap();
-        assert_eq!(value[0], PartialMoveChoice::SPECIFIC(0));
-        assert_eq!(value[1], PartialMoveChoice::SPECIFIC(1));
-
-        let value = iter.next().unwrap();
-        assert_eq!(value[0], PartialMoveChoice::SPECIFIC(1));
-        assert_eq!(value[1], PartialMoveChoice::SPECIFIC(1));
-
-        let value = iter.next().unwrap();
-        assert_eq!(value[0], PartialMoveChoice::SPECIFIC(2));
-        assert_eq!(value[1], PartialMoveChoice::SPECIFIC(1));
-
-        let value = iter.next().unwrap();
-        assert_eq!(value[0], PartialMoveChoice::SPECIFIC(0));
-        assert_eq!(value[1], PartialMoveChoice::SPECIFIC(2));
-
-        let value = iter.next().unwrap();
-        assert_eq!(value[0], PartialMoveChoice::SPECIFIC(1));
-        assert_eq!(value[1], PartialMoveChoice::SPECIFIC(2));
-
-        let value = iter.next().unwrap();
-        assert_eq!(value[0], PartialMoveChoice::SPECIFIC(2));
-        assert_eq!(value[1], PartialMoveChoice::SPECIFIC(2));
-
-        assert_eq!(iter.next(), None);
-    }
-
-    #[test]
-    fn delta_iterator_01() {
-        // player 0
-        let transitions = DynVec::NEST(vec![
-            // player 1
-            Arc::new(DynVec::NEST(vec![
-                // Player 2
-                Arc::new(DynVec::NEST(vec![
-                    // player 3
-                    Arc::new(DynVec::NEST(vec![
-                        // Player 4
-                        Arc::new(DynVec::NEST(vec![Arc::new(DynVec::BASE(1))])),
-                        // Player 4
-                        Arc::new(DynVec::NEST(vec![Arc::new(DynVec::BASE(2))])),
-                        // Player 4
-                        Arc::new(DynVec::NEST(vec![Arc::new(DynVec::BASE(3))])),
-                    ])),
-                ])),
-                // Player 2
-                Arc::new(DynVec::NEST(vec![
-                    // player 3
-                    Arc::new(DynVec::NEST(vec![
-                        // Player 4
-                        Arc::new(DynVec::NEST(vec![Arc::new(DynVec::BASE(4))])),
-                        // Player 4
-                        Arc::new(DynVec::NEST(vec![Arc::new(DynVec::BASE(5))])),
-                        // Player 4
-                        Arc::new(DynVec::NEST(vec![Arc::new(DynVec::BASE(1))])),
-                    ])),
-                ])),
-            ])),
-        ]);
-        let game_structure = EagerGameStructure {
-            player_count: 5,
-            labeling: vec![],
-            transitions: vec![transitions],
-            moves: vec![],
-        };
-        let state = 0;
-        let partial_move = vec![
-            PartialMoveChoice::SPECIFIC(0), // player 0
-            PartialMoveChoice::RANGE(2),    // player 1
-            PartialMoveChoice::SPECIFIC(0), // player 2
-            PartialMoveChoice::RANGE(3),    // player 3
-            PartialMoveChoice::SPECIFIC(0), // player 4
-        ];
-        let mut iter = DeltaIterator::new(&game_structure, state, &partial_move);
-
-        let value = iter.next().unwrap();
-        assert_eq!(value, 1);
-
-        let value = iter.next().unwrap();
-        assert_eq!(value, 2);
-
-        let value = iter.next().unwrap();
-        assert_eq!(value, 3);
-
-        let value = iter.next().unwrap();
-        assert_eq!(value, 4);
-
-        let value = iter.next().unwrap();
-        assert_eq!(value, 5);
-
-        // repeats state 1 again, but that is suppressed due to deduplication of emitted states
-
-        let value = iter.next();
-        assert_eq!(value, None);
-    }
-}
