+++ benches/benchmark_solver.rs
@@ -1,7 +1,7 @@
 use atl_checker::algorithms::certain_zero::distributed_certain_zero;
 use atl_checker::algorithms::certain_zero::search_strategy::bfs::BreadthFirstSearchBuilder;
 use atl_checker::atl::Phi;
-use atl_checker::edg::{ATLDependencyGraph, ATLVertex};
+use atl_checker::edg::atledg::{vertex::ATLVertex, ATLDependencyGraph};
 use atl_checker::game_structure::lcgs::ir::intermediate::IntermediateLCGS;
 use atl_checker::game_structure::lcgs::parse::parse_lcgs;
 use atl_checker::game_structure::EagerGameStructure;
